Make sure the type accessed by the LONG and ZLONG ASN1 type is really a long, to avoid problems on platforms where sizeof(int) != sizeof(long).

+++ crypto/dsa/dsa.h
@@ -117,7 +117,7 @@ struct dsa_st
 	/* This first variable is used to pick up errors where
 	 * a DSA is passed instead of of a EVP_PKEY */
 	int pad;
-	int version;
+	long version;
 	int write_params;
 	BIGNUM *p;
 	BIGNUM *q;	/* == 20 */
